refactor: move func to corresponding file
